fix(NWS8-04): localize formatPublicationDate to long month with active UI language
Change month:'short' to month:'long' in both bare-date and full-timestamp
branches of formatPublicationDate. Pass i18n.language.split('-')[0] from
NewsCard and NewsReaderSheet callers. Update test mocks to expose i18n
object and fix duplicate-text collisions caused by the new sr-only SheetTitle
added in NWS8-05.

+++ learn-greek-easy-frontend/src/utils/newsFormat.ts
@@ -33,7 +33,7 @@ export function formatPublicationDate(raw: string | null | undefined, locale?: s
     const [, y, mo, d] = m;
     return new Intl.DateTimeFormat(locale, {
       year: 'numeric',
-      month: 'short',
+      month: 'long',
       day: 'numeric',
       timeZone: 'UTC',
     }).format(new Date(Date.UTC(Number(y), Number(mo) - 1, Number(d))));
@@ -43,7 +43,7 @@ export function formatPublicationDate(raw: string | null | undefined, locale?: s
   const parsed = new Date(s);
   return Number.isNaN(parsed.getTime())
     ? ''
-    : parsed.toLocaleDateString(locale, { year: 'numeric', month: 'short', day: 'numeric' });
+    : parsed.toLocaleDateString(locale, { year: 'numeric', month: 'long', day: 'numeric' });
 }
 
 /**
